Move injectStyles and helpers to lib/plotcss_utils
This keeps the css directory free of Javascript. The helpers have been
moved inside the plotcss_utils.js file instead of a separate helpers
file. The plot API and tests have been updated accordingly.

Removed duplicate tests in plot_interact_tests.js
