+++ src/plot_api/plot_api.js
@@ -18,8 +18,6 @@ var Lib = require('../lib');
 var Events = require('../lib/events');
 var Queue = require('../lib/queue');
 
-var injectStyles = require('../css/plotcss_injector');
-
 var Plots = require('../plots/plots');
 var Fx = require('../plots/cartesian/graph_interact');
 
@@ -62,7 +60,7 @@ Plotly.plot = function(gd, data, layout, config) {
 
     // Inject the plot styles into the document where we're plotting, bails if
     // already styled
-    injectStyles(gd);
+    Lib.injectStyles(gd);
 
     // Events.init is idempotent and bails early if gd has already been init'd
     Events.init(gd);
